modify create_env.sh add parameters for all_clean realize is or not keep file create_env.sh exits:0 keep,other delete

+++ vim_scripts/create_env.sh
@@ -96,7 +96,11 @@ function all_clean()
 
     luf_clean    
 
-    echo "------------------------------------------------"
-    echo "cll_clean: lean all successed!"
-    rm -f create_env.sh
+    if [ $1 == 0 ]; then
+        echo "------------------------------------------------"
+        echo "cll_clean: lean all successed!"
+        rm -f create_env.sh
+    else
+        echo "Keep the file create_env.sh exists"
+    fi
 }
